LayoutTests/platform/chromium-win/plugins/mouse-events-fixedpos-expected.txt

CONSOLE MESSAGE: line 39: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 39: PLUGIN: mouseDown at (90, 90)
CONSOLE MESSAGE: line 40: PLUGIN: mouseUp at (90, 90)
CONSOLE MESSAGE: line 42: PLUGIN: mouseDown at (100, 100)
CONSOLE MESSAGE: line 44: PLUGIN: mouseUp at (110, 100)

Tests for widget positions being correctly updated after scrolling. rdar://problem/7559069
